requirements.txt
Flask~=3.0.3
flask-cors~=5.0.1
google-generativeai
python-dotenv~=1.0.1
PyMuPDF
pytesseract~=0.3.13
opencv-python~=4.11.0.86
Pillow
streamlit~=1.44.1

Werkzeug~=3.0.3
protobuf~=5.29.3
PyPDF2~=3.0.1
requests~=2.32.3